+++ packages/atomic/cypress/integration/result-templates/result-number-selectors.ts
@@ -1,0 +1,8 @@
+export const ResultNumberSelectors = {
+  component: 'atomic-result-number',
+  formats: {
+    unitFormat: 'atomic-format-unit',
+    currencyFormat: 'atomic-format-currency',
+    numberFormat: 'atomic-format-number',
+  },
+};