+++ packages/atomic/cypress/integration/search-interface.cypress.ts
@@ -1,27 +1,7 @@
 import {RouteAlias, setUpPage} from '../utils/setupComponent';
-import {i18n} from 'i18next';
-
-type SearchInterface = HTMLElement & {
-  language: string;
-  i18n: i18n;
-  executeFirstSearch: () => void;
-};
+import {getSearchInterface, setLanguage} from './search-interface-utils';
 
 describe('Search Interface Component', () => {
-  const getSearchInterface = (
-    cb: (searchInterface: SearchInterface) => void
-  ) => {
-    cy.get('atomic-search-interface').then(($el) => {
-      cb($el.get(0) as SearchInterface);
-    });
-  };
-
-  const setLanguage = (lang: string) => {
-    getSearchInterface((searchInterface) => {
-      searchInterface.language = lang;
-    });
-  };
-
   const setTranslation = (lang: string, key: string, value: string) => {
     getSearchInterface((searchInterface) => {
       searchInterface.i18n.addResource(lang, 'translation', key, value);
@@ -30,7 +10,7 @@ describe('Search Interface Component', () => {
 
   describe('with an automatic search', () => {
     beforeEach(() => {
-      setUpPage('<atomic-query-summary>', true);
+      setUpPage('<atomic-query-summary enable-duration>', true);
     });
 
     it('should support changing language', () => {