+++ packages/quantic/force-app/main/default/lwc/quanticCardContainer/quanticCardContainer.js
@@ -1,0 +1,8 @@
+import { LightningElement, api } from 'lwc';
+
+export default class QuanticCardContainer extends LightningElement {
+    /**
+     * @type {string}
+     */
+    @api title;
+}